fix(frontend/builder): split FormRenderer stories and address review feedback
Split 937-line FormRenderer.stories.tsx into category-based files:
- PrimitiveFields.stories.tsx (string, number, boolean, enum, date/time)
- ArrayFields.stories.tsx (list[str], list[int], list[Enum], list[bool], list[object])
- ComplexSchemas.stories.tsx (objects, unions, oneOf, multi-select, kitchen sink)
- FormRendererStoryWrapper.tsx (shared wrapper component and decorator)

Also fixed formData fallback (e.formData ?? {}) for robustness when
form is cleared.
